+++ bin/core/git/getGitLogs.ts
@@ -7,32 +7,43 @@ import type { GitLogsType } from '../../../typings'
 /**
  * 获取日志
  *
- * @param config - GitLogsType
- * @returns {Array} true 返回列表
+ * @param option - 参数
+ * @param option.lastet - 限制显示最近多长时间的日志
+ * @param option.limit - 限制数量
+ * @param option.noMerges - 是否排除merge记录
+ * @param option.params - 附带参数
+ * @param option.keys - 需要返回的字段
+ * @param option.grep - 筛选关键词
+ * @param option.author - 筛选提交人
+ * @param option.branch - 要查询的分支
+ * @returns logsList - 返回列表
  */
-function getGitLogs(config: any = {}): GitLogsType[] {
+function getGitLogs(option: any = {}): GitLogsType[] {
     const {
         lastet,
         limit,
-        // branches,
         params = '',
         keys,
-        noMerges = false
-    } = config
+        noMerges = false,
+        grep,
+        author,
+        branch
+    } = option
     const formatter = new GitLogsFormatter()
     let argv = [
         'log',
+        branch ? branch : '',
         '--date-order',
         `--pretty=format:${formatter.getFormat(keys)}`
     ]
     if (limit) argv.push('-' + limit)
     if (lastet) argv = argv.concat(['--since', getSeconds(lastet)])
+    if (grep) argv = argv.concat(['--grep', grep])
+    if (author) argv = argv.concat(['--author', author])
     if (noMerges) argv.push('--no-merges')
-    // if (branches) argv = argv.concat(['--branches', branches])
-    argv = argv.concat(params.split(' '))
+    if (params) argv = argv.concat(params.split(' '))
     const { stdout } = spawnSync('git', argv)
-    const logList: GitLogsType[] = formatter.getLogs(stdout)
-    return logList
+    return formatter.getLogs(stdout)
 }
 
 module.exports = getGitLogs