+++ docs/changelog.md
@@ -6,9 +6,10 @@
 
 1. 改用性能更好的原生脚本执行器，替代shelljs的`sh.exec`
 2. 新增`undo`指令，用于撤回代码
-3. `undo`指令新增`--no-merges` `--limit` `--lastet`传参
-4. `log`指令增加`--no-merges`传参，用于排除merge记录
-5. `process.exit`替换`sh.exit`
+3. `cleanbranch`指令流程优化，预览之后可选择分支进行清理
+4. `undo`指令新增`--no-merges` `--limit` `--lastet`传参
+5. `log`指令增加`--no-merges`传参，用于排除merge记录
+6. `process.exit`替换`sh.exit`
 
 ### 优化